[FIX] sale: add lines on a confirmed sale order.
When using the editable list on the sale order form,
it was not possible to add a new line on a confirmed sale order,
while it should
(when `Allow to edit sales order` is chosen in the Sales settings).

The reason was that the fields `tax_id` and `product_id` were
set as readonly when the SO state was not draft.

When using the popup/dialog view, it worked by luck, because
the line `state`, which is a stored related field to the sale order state,
was not correctly computed, and was seen as `draft`, while it should
actually be `sale`, for a confirmed sale order.

We change the readonly condition of these fields,
which is now based on the fact wether or not
the line is already invoiced or not, which makes
more sense.

opw-651573