[FIX] sale: prevent the change of product of a SO line when there are procurements
When adding a line to a confirmed sale order,
the procurements are automatically created, with
this product.

Therefore, the product of the line should no longer be changed,
as otherwise, the product on the procurement would no longer
be the same product than on the sale order line.

To be able to change the product of a line on a
confirmed sale order, you must therefore first
delete all procurements.