[IMP] res_partner: default image and gravatar
Try to find the gravatar of partners based on their email. Better computation
of the default image.

The default image is not computed as a default anymore. Indeed due to the
color of the default avatar it is not possible to know if the user has changed
the partner iamge or not.

An onchange on the email field allow to search for the gravatar. If no image is
given when creating a partner, the values are updated with a default image
coming from the contact type, whether the partner is a company or has a parent
with an image.

Many thanks to Yti and Rim that are waiting for this task to use their fancy
gravatar.