[FIX] calendar: reminder notifications: event's handlers
By extending the Notification widget to create the CalendarNotification one,
we override the events key of Notification defining the handler when clicking
on the close cross. So when clicking on the cross, it doesn't close the
notification, and it even toggle the app switcher since the preventDefault
isn't performed.

Note that this should be directly handled by the webclient itself, and its
mechanism of extending widgets. This should be done in current master.