[FIX] sale_mrp: stock transfer with BoM using variants
`_bom_explode` expects a `product.product`,
not a `product.template`

```
@param product: Select a particular variant of the BoM.
If False use BoM without variants.
```

This `product` attribute is used in the method `_skip_bom_line`,
which get the `attribute_value_ids` of this `product`.
This field `attribute_value_ids` is available on
`product.product` only, not on `product.template`

opw-656061