[FIX] point_of_sale: remove reference to currencies for cash control
The mentioned available currencies are a leftover from the old
account.bank.statement.cashbox which would show a pre-defined list of
'currencies' (eg. .05, .10, .50, 1, 2, 5, 20, 50, ... for Euro). Since
accounting v9 this is no longer how it works, the wizard now just
consists of value, amount pairs that the user has to fill out.

opw-653111