[FIX] website_sale_digital: fixed attachment count
The read_group function bypasses some restrictions, and in this case
was counting the "image", "image_small" and "image_medium" attachments
of the products, thus resulting in wrong attachment counts.

I changed it to a search_count which is cleaner, simplifies the code
a lot and, more importantly, returns the right values.