[IMP] purchase: split onchange_product_id
This commit splits the changes which affect the product_id and the
changes which affect the product quantity. The main reason is that a
product change should trigger the recalculation of the price unit,
while a change in the quantity should not trigger product-specific data
recalculation (such as the name or the taxes).

The original issue was the following:
- set a supplier for a product with a minimum quantity larger than 1.0
- create a PO for this supplier and this product: since the minimum
  quantity is not met, no tax is set on the line.
- change the quantity to the minimum quantity of the supplier
- the tax is set on the line, but a warning pop up confuses the user by
  asking if it is necessary to discard the line since Odoo and the user
  changed the same line.

This commit prevents this warning, and moreover it will set the taxes on
the product right away.

Bonus: if you change the product of an existing line, it will not keep
the privious quantity/price values, which was the case before.

opw-656729