[FIX] mass_mailing: /unsubscribe controller must be auth='public'
Rendering a website template with auth='none' has as side-effect
to render the website template as sudo, therefore display "Administrator"
instead of "Sign in" when the user is not signed in in the website
user menu, as well as other issues, e.g. records browsed as sudo
while it's not required.