[FIX] mass_mailing: Remove crappy onchange
There was a misunderstanding in the model implementation for campaigns.
The inheritS was handled in a strange way while redefining the name field,
and creating the utm campaign_id field in an onchange.

A stable commit has been given in saas-6, while this one will remove the onchange in saas-7
What's done:
- Remove the required name field
- Set campaign_id as name with the _rec_name property
- Remove the onchange on the model and the view
- Everything will be handled thanks to the inheritS mechanism